requirements.txt
python-telegram-bot==20.6
gspread==5.12.0
oauth2client==4.1.3
